Implement fetching data by chunks, improved UX importing data, and speed after first load.

+++ lib/data-utils.ts
@@ -2,7 +2,13 @@ import * as Etebase from 'etebase';
 import Swal from 'sweetalert2';
 import moment from 'moment';
 
-import { sortByName, sortByDate, showNotification, doLogin } from './utils';
+import {
+  sortByName,
+  sortByDate,
+  showNotification,
+  doLogin,
+  uniqBy,
+} from './utils';
 import * as T from './types';
 
 const ETEBASE_SERVER_URL = process.env.NEXT_PUBLIC_ETEBASE_SERVER_URL;
@@ -15,6 +21,22 @@ const collectionTypes = {
 const collectionUids = {
   budgets: '',
   expenses: '',
+};
+
+// While this defeats the purpose of e2ee a little bit, the app is just too slow without it.
+const cachedData: { budgets: T.Budget[]; expenses: T.Expense[] } = {
+  budgets: [],
+  expenses: [],
+};
+
+const hasStartedLoading = {
+  budgets: false,
+  expenses: false,
+};
+
+const hasFinishedLoading = {
+  budgets: false,
+  expenses: false,
 };
 
 export const validateLogin = async (email: string, syncToken: string) => {
@@ -127,6 +149,76 @@ const getBudgetFromItem = async (item: Etebase.Item) => {
   }
 };
 
+const getExpenseFromItem = async (item: Etebase.Item) => {
+  try {
+    const data: T.ExpenseContent = JSON.parse(
+      await item.getContent(Etebase.OutputFormat.String),
+    );
+
+    return {
+      id: item.uid,
+      cost: data.cost,
+      description: data.description,
+      budget: data.budget,
+      date: data.date,
+    } as T.Expense;
+  } catch (error) {
+    return null;
+  }
+};
+
+const loadItemsAsync = async (
+  type: 'budgets' | 'expenses',
+  itemManager: Etebase.ItemManager,
+  stoken?: string,
+) => {
+  if (hasFinishedLoading[type] || (hasStartedLoading[type] && !stoken)) {
+    return cachedData[type];
+  }
+
+  const items = await itemManager.list({ limit: 500, stoken });
+
+  // TODO: Remove this
+  console.log('======== data-utils.loadItemsAsync -- items');
+  console.log(items);
+
+  if (type === 'budgets') {
+    hasStartedLoading.budgets = true;
+
+    const budgets = (
+      await Promise.all(items.data.map(getBudgetFromItem))
+    ).filter((budget) => Boolean(budget));
+
+    if (!items.done) {
+      // eslint-disable-next-line
+      loadItemsAsync(type, itemManager, items.stoken);
+    } else {
+      hasFinishedLoading.budgets = true;
+    }
+
+    cachedData.budgets = uniqBy([...cachedData.budgets, ...budgets], 'id');
+
+    return cachedData.budgets;
+  }
+
+  hasStartedLoading.expenses = true;
+
+  const expenses = (
+    await Promise.all(items.data.map(getExpenseFromItem))
+  ).filter((expense) => Boolean(expense));
+
+  if (!items.done) {
+    // eslint-disable-next-line
+    loadItemsAsync(type, itemManager, items.stoken);
+  } else {
+    hasFinishedLoading.expenses = true;
+  }
+
+  cachedData.expenses = uniqBy([...cachedData.expenses, ...expenses], 'id');
+
+  return cachedData.expenses;
+};
+
 export const fetchBudgets = async (
   etebase: Etebase.Account,
   month?: string,
@@ -135,14 +227,11 @@ export const fetchBudgets = async (
     const collectionManager = etebase.getCollectionManager();
     const collection = await collectionManager.fetch(collectionUids.budgets);
     const itemManager = collectionManager.getItemManager(collection);
-    const items = await itemManager.list();
 
-    const budgets = (await Promise.all(items.data.map(getBudgetFromItem)))
+    const budgets: T.Budget[] = (
+      (await loadItemsAsync('budgets', itemManager)) as T.Budget[]
+    )
       .filter((budget) => {
-        if (!budget) {
-          return false;
-        }
-
         if (!month) {
           return true;
         }
@@ -168,24 +257,6 @@ export const fetchBudgets = async (
   return [];
 };
 
-const getExpenseFromItem = async (item: Etebase.Item) => {
-  try {
-    const data: T.ExpenseContent = JSON.parse(
-      await item.getContent(Etebase.OutputFormat.String),
-    );
-
-    return {
-      id: item.uid,
-      cost: data.cost,
-      description: data.description,
-      budget: data.budget,
-      date: data.date,
-    } as T.Expense;
-  } catch (error) {
-    return null;
-  }
-};
-
 export const fetchExpenses = async (
   etebase: Etebase.Account,
   month?: string,
@@ -194,14 +265,11 @@ export const fetchExpenses = async (
     const collectionManager = etebase.getCollectionManager();
     const collection = await collectionManager.fetch(collectionUids.expenses);
     const itemManager = collectionManager.getItemManager(collection);
-    const items = await itemManager.list();
 
-    const expenses = (await Promise.all(items.data.map(getExpenseFromItem)))
+    const expenses = (
+      (await loadItemsAsync('expenses', itemManager)) as T.Expense[]
+    )
       .filter((expense) => {
-        if (!expense) {
-          return false;
-        }
-
         if (!month) {
           return true;
         }
@@ -272,8 +340,6 @@ export const saveBudget = async (
     const itemManager = collectionManager.getItemManager(collection);
 
     if (budget.id === 'newBudget') {
-      budget.id = `${Date.now().toString()}:${Math.random()}`;
-
       const item = await itemManager.create(
         {
           type: 'budget',
@@ -285,8 +351,16 @@ export const saveBudget = async (
           month: budget.month,
         } as T.BudgetContent),
       );
+      budget.id = item.uid;
 
       await itemManager.batch([item]);
+
+      cachedData.budgets.push({
+        id: budget.id,
+        name: budget.name,
+        value: budget.value,
+        month: budget.month,
+      });
     } else {
       const existingBudgetItem = await itemManager.fetch(budget.id);
       const existingBudget = await getBudgetFromItem(existingBudgetItem);
@@ -301,6 +375,14 @@ export const saveBudget = async (
         } as T.BudgetContent),
       );
       await itemManager.batch([existingBudgetItem]);
+
+      const cachedBudgetItemIndex = cachedData.budgets.findIndex(
+        (_budget) => _budget.id === budget.id,
+      );
+      if (cachedBudgetItemIndex !== -1) {
+        cachedData.budgets[cachedBudgetItemIndex].name = budget.name;
+        cachedData.budgets[cachedBudgetItemIndex].value = budget.value;
+      }
 
       // Update all expenses with the previous budget name to the new one, if it changed
       if (oldName !== newName) {
@@ -324,6 +406,13 @@ export const saveBudget = async (
             } as T.ExpenseContent),
           );
           expenseItemsToUpdate.push(expenseItem);
+
+          const cachedItemIndex = cachedData.expenses.findIndex(
+            (_expense) => _expense.id === expense.id,
+          );
+          if (cachedItemIndex !== -1) {
+            cachedData.expenses[cachedItemIndex].budget = newName;
+          }
         }
         await expensesItemManager.batch(expenseItemsToUpdate);
       }
@@ -376,6 +465,13 @@ export const deleteBudget = async (
 
     item.delete();
     await itemManager.batch([item]);
+
+    const cachedItemIndex = cachedData.budgets.findIndex(
+      (budget) => budget.id === budgetId,
+    );
+    if (cachedItemIndex !== -1) {
+      cachedData.budgets.splice(cachedItemIndex, 1);
+    }
 
     return true;
   } catch (error) {
@@ -450,6 +546,13 @@ export const saveExpense = async (
       );
 
       await itemManager.batch([item]);
+
+      cachedData.budgets.push({
+        id: item.uid,
+        name: expense.budget,
+        value: 100,
+        month: expense.date.substring(0, 7),
+      });
     }
 
     const collectionManager = etebase.getCollectionManager();
@@ -457,8 +560,6 @@ export const saveExpense = async (
     const itemManager = collectionManager.getItemManager(collection);
 
     if (expense.id === 'newExpense') {
-      expense.id = `${Date.now().toString()}:${Math.random()}`;
-
       const item = await itemManager.create(
         {
           type: 'expense',
@@ -473,6 +574,16 @@ export const saveExpense = async (
       );
 
       await itemManager.batch([item]);
+
+      expense.id = item.uid;
+
+      cachedData.expenses.push({
+        id: expense.id,
+        cost: expense.cost,
+        budget: expense.budget,
+        description: expense.description,
+        date: expense.date,
+      });
     } else {
       const existingExpenseItem = await itemManager.fetch(expense.id);
 
@@ -485,6 +596,16 @@ export const saveExpense = async (
         } as T.ExpenseContent),
       );
       await itemManager.batch([existingExpenseItem]);
+
+      const cachedItemIndex = cachedData.expenses.findIndex(
+        (_expense) => _expense.id === expense.id,
+      );
+      if (cachedItemIndex !== -1) {
+        cachedData.expenses[cachedItemIndex].cost = expense.cost;
+        cachedData.expenses[cachedItemIndex].description = expense.description;
+        cachedData.expenses[cachedItemIndex].budget = expense.budget;
+        cachedData.expenses[cachedItemIndex].date = expense.date;
+      }
     }
 
     return true;
@@ -513,6 +634,13 @@ export const deleteExpense = async (
     item.delete();
     await itemManager.batch([item]);
 
+    const cachedItemIndex = cachedData.expenses.findIndex(
+      (expense) => expense.id === expenseId,
+    );
+    if (cachedItemIndex !== -1) {
+      cachedData.expenses.splice(cachedItemIndex, 1);
+    }
+
     return true;
   } catch (error) {
     Swal.fire({
@@ -528,15 +656,48 @@ export const deleteExpense = async (
 
 export const deleteAllData = async (etebase: Etebase.Account) => {
   const collectionManager = etebase.getCollectionManager();
-  const collections = await collectionManager.list([
-    collectionTypes.budget,
-    collectionTypes.expense,
-  ]);
+  const budgetsCollection = await collectionManager.fetch(
+    collectionUids.budgets,
+  );
+  const budgetsItemManager =
+    collectionManager.getItemManager(budgetsCollection);
 
-  for (const collection of collections.data) {
-    collection.delete();
-    await collectionManager.upload(collection);
+  const budgets = await fetchBudgets(etebase);
+  const budgetItems: Etebase.Item[] = [];
+  for (const budget of budgets) {
+    const item = await budgetsItemManager.fetch(budget.id);
+    item.delete();
+    budgetItems.push(item);
   }
+  await budgetsItemManager.batch(budgetItems);
+
+  const expensesCollection = await collectionManager.fetch(
+    collectionUids.expenses,
+  );
+  const expensesItemManager =
+    collectionManager.getItemManager(expensesCollection);
+
+  const expenses = await fetchExpenses(etebase);
+  const expenseItems: Etebase.Item[] = [];
+  for (const expense of expenses) {
+    const item = await expensesItemManager.fetch(expense.id);
+    item.delete();
+    expenseItems.push(item);
+  }
+  await expensesItemManager.batch(expenseItems);
+
+  budgetsCollection.delete();
+  await collectionManager.upload(budgetsCollection);
+
+  expensesCollection.delete();
+  await collectionManager.upload(expensesCollection);
+
+  cachedData.budgets.length = 0;
+  cachedData.expenses.length = 0;
+  hasStartedLoading.budgets = false;
+  hasStartedLoading.expenses = false;
+  hasFinishedLoading.budgets = false;
+  hasFinishedLoading.expenses = false;
 };
 
 type ExportAllData = (etebase: Etebase.Account) => Promise<{
@@ -547,8 +708,12 @@ type ExportAllData = (etebase: Etebase.Account) => Promise<{
 export const exportAllData: ExportAllData = async (
   etebase: Etebase.Account,
 ) => {
+  // Don't import anything until we're done with the first full load
+  if (!hasFinishedLoading.budgets || !hasFinishedLoading.expenses) {
+    return {};
+  }
+
   try {
-    // NOTE: The queries look weird because .dump() and simple .find() were returning indexes and other stuff
     const budgets = (await fetchBudgets(etebase)).sort(sortByName);
     const expenses = (await fetchExpenses(etebase)).sort(sortByDate);
 
@@ -572,6 +737,11 @@ export const importData = async (
   budgets: T.Budget[],
   expenses: T.Expense[],
 ) => {
+  // Don't import anything until we're done with the first full load
+  if (!hasFinishedLoading.budgets || !hasFinishedLoading.expenses) {
+    return false;
+  }
+
   try {
     if (replaceData) {
       await deleteAllData(etebase);
@@ -601,6 +771,13 @@ export const importData = async (
       );
 
       budgetItems.push(item);
+
+      cachedData.budgets.push({
+        id: item.uid,
+        name: budget.name,
+        value: budget.value,
+        month: budget.month,
+      });
     }
 
     await budgetsItemManager.batch(budgetItems);
@@ -627,6 +804,14 @@ export const importData = async (
       );
 
       expenseItems.push(item);
+
+      cachedData.expenses.push({
+        id: item.uid,
+        cost: expense.cost,
+        budget: expense.budget,
+        description: expense.description,
+        date: expense.date,
+      });
     }
 
     await expensesItemManager.batch(expenseItems);
@@ -649,6 +834,11 @@ export const copyBudgets = async (
   originalMonth: string,
   destinationMonth: string,
 ) => {
+  // Don't copy anything until we're done with the first full load
+  if (!hasFinishedLoading.budgets || !hasFinishedLoading.expenses) {
+    return;
+  }
+
   const originalBudgets = await fetchBudgets(etebase, originalMonth);
   const destinationBudgets = originalBudgets.map((budget) => {
     const newBudget: T.Budget = { ...budget };
@@ -677,6 +867,13 @@ export const copyBudgets = async (
         );
 
         items.push(item);
+
+        cachedData.budgets.push({
+          id: item.uid,
+          name: budget.name,
+          value: budget.value,
+          month: budget.month,
+        });
       }
 
       await itemManager.batch(items);
